Refactor storage_deposit and byte_cost_config modules
