+++ src/core/CMakeLists.txt
@@ -17,6 +17,7 @@ target_sources(
           "utils/uint256.c"
           "utils/macros.h"
           "models/inputs/utxo_input.c"
+          "models/outputs/byte_cost_config.c"
           "models/outputs/feat_blocks.c"
           "models/outputs/output_alias.c"
           "models/outputs/output_basic.c"
@@ -32,7 +33,6 @@ target_sources(
           "models/message.c"
           "models/unlock_block.c"
   PUBLIC "address.h"
-         "byte_cost_config.h"
          "utils/iota_str.h"
          "utils/byte_buffer.h"
          "utils/slip10.h"
@@ -40,6 +40,7 @@ target_sources(
          "utils/uint256.h"
          "utils/macros.h"
          "models/inputs/utxo_input.h"
+         "models/outputs/byte_cost_config.h"
          "models/outputs/feat_blocks.h"
          "models/outputs/output_alias.h"
          "models/outputs/output_basic.h"
