+++ src/core/models/outputs/storage_deposit.c
@@ -1,18 +1,7 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#include <inttypes.h>
-
-#include "core/models/inputs/utxo_input.h"
-#include "core/models/message.h"
 #include "core/models/outputs/storage_deposit.h"
-
-// Defines the rent of a single virtual byte denoted in IOTA tokens
-#define DEFAULT_BYTE_COST 500
-// Defines the multiplier for data fields
-#define DEFAULT_BYTE_COST_FACTOR_DATA 1
-// Defines the multiplier for fields which can act as keys for lookups
-#define DEFAULT_BYTE_COST_FACTOR_KEY 10
 
 static uint64_t calc_minimum_output_deposit(byte_cost_config_t *config, utxo_output_type_t output_type, void *output) {
   if (config == NULL || output == NULL) {
@@ -61,31 +50,7 @@ static uint64_t basic_address_storage_deposit(byte_cost_config_t *config, addres
   return config->v_byte_cost * ((output_serialized_len * config->v_byte_factor_data) + config->v_byte_offset);
 }
 
-byte_cost_config_t *storage_deposit_new_config(uint16_t byte_cost, uint8_t byte_factor_data, uint8_t byte_factor_key) {
-  byte_cost_config_t *config = malloc(sizeof(byte_cost_config_t));
-  if (!config) {
-    printf("[%s:%d] can not create storage config\n", __func__, __LINE__);
-    return NULL;
-  }
-
-  config->v_byte_cost = byte_cost;
-  config->v_byte_factor_data = byte_factor_data;
-  config->v_byte_factor_key = byte_factor_key;
-
-  // size of: output ID + message ID + confirmation milestone index + confirmation unix timestamp
-  config->v_byte_offset = (IOTA_OUTPUT_ID_BYTES * byte_factor_key) +    // output ID
-                          (IOTA_MESSAGE_ID_BYTES * byte_factor_data) +  // message ID
-                          (sizeof(uint32_t) * byte_factor_data) +       // confirmation milestone index
-                          (sizeof(uint32_t) * byte_factor_data);        // confirmation unix timestamp
-
-  return config;
-}
-
-byte_cost_config_t *storage_deposit_new_default_config() {
-  return storage_deposit_new_config(DEFAULT_BYTE_COST, DEFAULT_BYTE_COST_FACTOR_DATA, DEFAULT_BYTE_COST_FACTOR_KEY);
-}
-
-bool storage_deposit_check_sufficient_output_deposit(byte_cost_config_t *config, utxo_output_type_t output_type,
+bool storage_deposit_sufficient_output_deposit_check(byte_cost_config_t *config, utxo_output_type_t output_type,
                                                      void *output) {
   if (config == NULL || output == NULL) {
     printf("[%s:%d] invalid parameters\n", __func__, __LINE__);
