+++ src/core/models/outputs/storage_deposit.h
@@ -4,25 +4,12 @@
 #ifndef __CORE_MODELS_OUTPUT_STORAGE_DEPOSIT_H__
 #define __CORE_MODELS_OUTPUT_STORAGE_DEPOSIT_H__
 
-#include "core/byte_cost_config.h"
+#include "core/models/outputs/byte_cost_config.h"
 #include "core/models/outputs/outputs.h"
 
-/**
- * @brief Create new byte cost configuration
- *
- * @param[in] byte_cost Rent of a single virtual byte denoted in IOTA tokens
- * @param[in] byte_factor_data Multiplier for data fields
- * @param[in] byte_factor_key Multiplier for key fields
- * @return *byte_cost_config_t
- */
-byte_cost_config_t *storage_deposit_new_config(uint16_t byte_cost, uint8_t byte_factor_data, uint8_t byte_factor_key);
-
-/**
- * @brief Create new default byte cost configuration
- *
- * @return *byte_cost_config_t
- */
-byte_cost_config_t *storage_deposit_new_default_config();
+#ifdef __cplusplus
+extern "C" {
+#endif
 
 /**
  * @brief Check if a sufficient storage deposit was made for the given output
@@ -32,7 +19,11 @@ byte_cost_config_t *storage_deposit_new_default_config();
  * @param[in] output Pointer to an output
  * @return true if output has enough storage deposit amount
  */
-bool storage_deposit_check_sufficient_output_deposit(byte_cost_config_t *config, utxo_output_type_t output_type,
+bool storage_deposit_sufficient_output_deposit_check(byte_cost_config_t *config, utxo_output_type_t output_type,
                                                      void *output);
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif  // __CORE_MODELS_OUTPUT_STORAGE_DEPOSIT_H__
